+++ tests/testthat.R
@@ -1,6 +1,6 @@
 library(testthat)
 library(reportRx)
 test_check("reportRx")
-# NOTE: Do not alter this file. Testing sripts should be placed in the testthat folder.
+# NOTE: Do not alter this file. Testing scripts should be placed in the testthat folder.